+++ src/components/bookedList/BookedWeb.tsx
@@ -27,22 +27,19 @@ const BookedList = () => {
   return (
     <>
       {localHotelData === null ? (
-        <div className="flex flex-col h-full">
-          <div className="flex items-center justify-center w-full h-16 py-2 bg-white border-b-4 border-slate-200">
-            예약된 내역
-          </div>
-          <div className="flex w-full h-16 py-2 bg-white rounded cursor-pointer text-slate-500">
-            <div className="flex items-center justify-center w-1/3 border-b-2 border-black">
+        <div className="flex h-full">
+          <div className="flex flex-col w-1/4 h-40 py-2 bg-white rounded cursor-pointer">
+            <div className="flex items-center w-full h-1/3 border-l-4 border-[#FF375C] bg-[#FEEEF1] pl-6 ">
               예정된 예약
             </div>
-            <div className="flex justify-center items-center w-1/3  bl-[#FF375C]">
+            <div className="flex items-center w-full h-1/3 bl-[#FF375C] pl-6">
               취소된 예약
             </div>
-            <div className="flex justify-center items-center w-1/3  bl-[#FF375C]">
+            <div className="flex items-center w-full h-1/3 bl-[#FF375C] pl-6">
               투숙 완료
             </div>
           </div>
-          <div className="flex items-center justify-center w-full py-2 ml-4 bg-white h-[60vh]">
+          <div className="flex items-center justify-center w-3/4 py-2 ml-4 bg-white h-[60vh]">
             <div className="flex flex-col items-center text-center">
               <img src={Noreserve} className="w-20 mb-5" />
               아직 준비된 예약이 없어요. <br />
